amazon-ppc-fixed/requirements.txt

# Amazon PPC Optimizer - Python Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
requests>=2.28.0
pyyaml>=6.0
functions-framework>=3.0.0

# Optional but recommended
python-dateutil>=2.8.2
pytz>=2023.3

# For enhanced logging and reporting
colorama>=0.4.6

# For BigQuery audit logging
google-cloud-bigquery>=3.11.0

# For data analysis (optional)
# pandas>=1.5.0
# numpy>=1.24.0
